command parser syntax
Since styles should not have a dot in them we can now use them to
specify a language. For instance when converting to markdown we can
use a style converter (cstyle). Then to write we can specify say
`xml._` to say that after we convert the file we want to write it as
an xml language in the default style. This is particularly useful when
trying to view the whole document.