Setting up __main__
The command line utility will now be able to issue commands and
redirect the command to a module. The __main__ file should not need to
be modified any longer. If there is a problem, then it will be with
one of the commands. As of now, the config, defaults and edit commands
work.