lexor 0.0.8b0
This version seems handle well the command line arguments for each of
the styles. Note how each of the processors makes the call to
`config.set_style_cfg`, this call reads the configuration file to
obtain the data stored for a particular command. There may be a way of
reducing the number of calls to reading the configuration file since
it is not expected to change so much while parsing, converting and
writer but for now this is not an issue. For now development needs to
be done for the cpp parser.