Writer BUG
Wrote a comment pointing out a possible fix. The problem we are facing
is that when we overload the write_child function then the write_end
does not get called. The current solution in the styles is to call the
function end if we want it. This however, does not look so great.
Check this next time.