Document defaults attribute
The `__str__` function allows the print function to display a `node`.
This however prevents us from displaying the node with different
options. These options can be controlled with the configuration files
or the command line. To be able to use the options I have added the
attribute `defaults` to the `Document` object. This will be mainly use
to overwrite the defaults attribute in the `Writer` object. The other
`Parser` object and the `Converter` object defaults can be easily
accessed and thus the defaults attribute for the Document will not be
used.