lexor 0.0.8b9
The language module now has the same naming convention to styles
across all functions. The test module has the addition of the
`nose_msg_explanations`. This neat little function reads the
`MSG_EXPLANATION` dictionary from each node parser and runs the tests
embedded in them.  The `repr` function in the elements module had a
slight mistake when writing log nodes. This has been fixed. Since the
test function now runs all the tests there is no need to write a
separate test for each case. You may write more tests if necessary,
but as far as the parser testing goes this is enough.