Install from web
If a style is not found in the local disk then it searches the github
website for possible matches and installs them. This is currently a
quick hack to be able to share the styles. This needs more thought to
make it behave similar to easy_install, pip and other installation
tools. The writer module now provides the function `get_node_writer`.
This is a way of accessing the other `NodeWriter`s available to the
Writer.