to command fixes
lexor will no longer crash if the xml module is not installed. This
was fixed by setting the node converters when the converter converts a
style. A switch was put in place similar as the writer. If we wish to
set some parameters for the log then we no longer need to state the
log style. By default we use the lexor:log style, we just need to
introduce parameters as follows: `--log @module=on@explanation=on`. In
the future, the log style will have other options (html) so that we
don't have to create many log styles.