v0.0.9b5
Updated the `__len__` function in the `Node` object. The writer object
now gives us a way of knowing the last string it wrote: `prev_str`.
All the NodeWriters can now communicate with each other if we use
references, that is if we are using a node writer for say `#text` and
we wish to use the same one for `#entity` then instead of issuing a
new node writing for `#entity` we simply say `#text`.