parser and converter messages
A few updates to the parser and converter messages. They are now
compatible with the new message system. In the parser I deleted the
deprecated functions since all the new style modules are now being
hosted at github and they seem to be pretty stable now. There was a
mistake in the elements module. This mistake happened when we extended
a list. This has been fixed. Since we are moving on to the converting
styles it is time to spend some time adding more methods to the `Node`
objects.