v0.0.9b8
The converter now has the function `exec_python`. This is useful for
converting styles that may handle python code. The inserting functions
have been added to elements. The writer module had some bugs fixed. In
particular, if the string being printed is empty then there we do
nothing. The other bug was in the write function which would call the
write_end even when the child said it would take care of it. The to
command had other bugs that have been fixed too.