writer wrapper
The writer object now has the function `wrap`. This function makes use
of the `buffer` attribute to be able to print characters within a
given width. Note that the function `write` and `wrap` should be used
in the same style. If you are planning on wrapping everything at a
given width then use the function wrap (this is somewhat expensive due
to the use of the buffer), otherwise only use the function write. See
the html default style to see how to use this new function.