exec_python
A few more modifications done to it. In particular we now have the
"magic"  keywords `__FILE__`, `__DIR__` and `__NODE__`. These
variables gives us the full path to the calling file. The directory
and the node in which the code is being executed. We also have the
function `import_module` which allows you to import a module relative
to the location of the file executing.