raw, wrap, indent
The wrap method does not exists any longer. From now on the `write`
method in the NodeWriter will do all the work. To be able to control
the writer we have the `enable_raw`, `enable_wrap` methods. We can
control the indent also but this is the scope of another update. See
the HTML default writer.