writer updates
A few updates. Line 271 shows the lack of understanding of a simple
function. By fixing this the rest becomes more simple. The only new
thing to add here is to check if `end` is the same as the length of
the line. If this is the case then we need to wait until the buffer
has grown more and there is a place where we can break it. Otherwise
we will only be introducing new spaces.