writer updates: indent
Added the indent property so that we may change the indentation. This
forces a buffer flush with no tail. This means that if need the extra
white space after changing the indentation you must call the
`flush_buffer` manually. I may add a set function which allows you to
specify the tail property in flush_buffer later on.