v0.1c7
The converter no longer needs the `init_converter` function. If it is
declared then it will be used, but otherwise you can skip creating it.
You do however need to create the `convert` function. This may go away
in the future since I am still new to creating converters and I really
don't know any better right now. There is a function deprecation in
the writer module. The parser as well as the writer and parser can now
takes string values in the MAPPING dictionary. This is done so that
the object may avoid creating different instances of a node-
parser/writer/converter.