v0.1.1c0
version 0.1.0c8 and 0.1.0c9 have been skipped since quite a few
changes took place. Very minor change in test (bug). In the elements
module I added the function rename and added an accessor for the
attribute name. Some stuff was moved around but nothing major. The
`to` command needed some modifications since the converter module took
some major changes. The convert function now returns the doc and the
log. Do not access the document and log with `converter.doc` and
`converter.log`, these are now lists. Now we can move on to the
converter. The node converters got rid of the process method and now
we have the start and end method. The start method is called on the
first encounter with a node. If the node has children you will have
another encounter with it, this is when the end method gets called.
The comments are poorly written but they should be sufficient for now.
Namespaces, there has to be a section on this. All the converters
access only one namespace, be careful how you use this. Also, each
document has a namespace. More modifications will take place but these
will be incremental from now on. That's it, hopefully the first
converter will be ready sometime today.