import errors
An import error will be raised by the `lang` module when it tries to
find a style but it fails. This is now caught by the command `to`. If
lexor can't find your module then you should check to see if you can
import the module manually into python. Chances are there is a bug in
the module and lexor will not be able to import it.