get_element_by_id
This method has been added to the Document object. The implementation
is done via a dictionary which is updated by the node `set_parent` and
`disconnect`. In this way the correct id (if it exists) will always be
set to the document and there will be no need to look up traverse the
object. Note however, that there is a potential crash if two elements
have the same id. I will worry about that on later tests. The
converter method _get_direction got reduced and I started working on
the Selector object which will behave similarly to jquery.