modules organization
The Node object has been moved to its own file. Due to this I had to
call the elements module by accessing the sys.modules dictionary. The
import notation has been changed on all the lexor modules. No new
changes have been made, only notation, with the exception of the
writer module which now does not check to see if an element is a
character data. Instead it now checks if it has a data attribute. So
make sure that no other elements have the data attribute besides a
character data.