v0.1.1c10
Would have gone to 0.1.2 but this is still buggy. The selector can now
append and prepend. The node is missing the function contains and thus
the selector does not give us the correct id element. A document is
set to be temporary so that if is ever appended somewhere we may know
if we want to adopt its nodes or put the whole document.