element attributes!!! BUG FIX
SUPER BIG NOTE: If you ever want to attach an attribute to a node,
this is something you are doing a lot actually, you will find more
bugs like this in the future. Refer to this, when you delete a
property from the node you are also removing a possible element
attribute since they both share the same dictionary. Append `__` to
the name instead so that you make a distinction between element
attributes and properties. A few other changes were made to fix other
bugs.