keep your sh*t together Manuel
The clone node method should not be deep, you must specify if you want
a deep copy. In the converter you are traversing through the whole
tree so you DO NOT want a keep copy. Also, the clone_node in the
elements was not done correctly. This fix seems to have it all. It
correctly copies the Document with the deep mode on and off.