Github Move
All the changes that have been done to lexor since the first move
will continue in Github. There are some language styles that have not
been updated yet to conform to the latest changes to the source code
of lexor. The next commits will perhaps be hit or misses to see how I
can move everything to the development branch.