[BUG FIX]: converter
There was more progress in the documentation. The converter convert
function now sets the correct language, that is, the language to
which it was converted. This was not an issue before because I always
used the Writer and told it explicitly what language to use when
printing. The problem was brought to light when I converted and then
tried to use the ``__str__`` method on the Document.