?py ideas
Since the compile built in function throws only certain errors we can
compile the the text during the parsing phase, this way we won’t have
to parse the code each time during the conversion phase. Also, if we do
choose to do this we can pinpoint the exact place where the python
errors happen. Each processing instruction node will have a copy of its
compiled code with the information of the file name and its correct
line number. Time to go to the parsing side now, well, first add this
functionality to the ProcessingInstruction Node and then use it in the
parsers. Good luck.